feat: implements gzip compression caching middleware
