+++ src/server/index.js
@@ -10,7 +10,7 @@ const routes = require('./routes');
 require('dotenv').config();
 const language = require('./language');
 
-const cacheMiddleware = require('./middlewares/cacheMiddleware');
+const cacheMiddleware = require('./middlewares/cacheZipMiddleware');
 
 const app = express();
 const port = process.env.NODE_PORT || 8080;
@@ -48,7 +48,7 @@ function start() {
     .use(bodyParser.urlencoded({ extended: false }))
     .use(bodyParser.json())
     .use(language)
-    .use(cacheMiddleware)
+    .use(cacheGZipMiddleware)
     .use('/', routes)
     .use((_req, res) => res.status(404).json({ success: false, error: 'Route not found' }));
 }
